ImageBinding Adapter 정리
